added released filter for questions in exam and added delete button to exam view

+++ questionbank_extensions/remove_from_quiz.php
@@ -57,8 +57,12 @@ class remove_from_quiz extends column_base {
                                               JOIN {quiz_slots} qs ON qr.itemid = qs.id
                                               WHERE qr.component='mod_quiz' AND qr.questionarea = 'slot' AND qs.quizid = ? AND qr.questionbankentryid = ?", array($quizid, $question->questionbankentryid));
 
-
-        echo '<button href="#" id="removequestion' . $question->questionbankentryid.'" class="removequestion' . $question->questionbankentryid . ' btn btn-primary" type="button" value="removequestion"> ' . get_string('remove_from_quiz', 'block_exaquest') . '</button>';
+        //check if already in the quiz
+        if($DB->record_exists_sql("SELECT *
+                                    FROM {question_references} qr
+                                         JOIN {quiz_slots} qs ON qr.itemid = qs.id
+                                   WHERE qr.component='mod_quiz' AND qr.questionarea = 'slot' AND qs.quizid = ? AND qr.questionbankentryid = ?", array($quizid, $question->questionbankentryid))) {
+            echo '<button href="#" id="removequestion' . $question->questionbankentryid . '" class="removequestion' . $question->questionbankentryid . ' btn btn-primary" type="button" value="removequestion"> ' . get_string('remove_from_quiz', 'block_exaquest') . '</button>';
 
 
 
@@ -99,6 +103,10 @@ class remove_from_quiz extends column_base {
         </script>
         <?php
 
-
+        }
+    }
+    public function get_extra_joins(): array {
+        return ['qs' => 'JOIN {block_exaquestquestionstatus} qs ON qbe.id = qs.questionbankentryid',
+            'qra' => 'LEFT JOIN {block_exaquestreviewassign} qra ON qbe.id = qra.questionbankentryid'];
     }
 }
